setThrowPositionTimely
