+++ src/physics.js
@@ -5,44 +5,30 @@ import entityInfo from "./entities/entitiesInfo";
 import adjustDegree from "./utils/physicsUtils/adjustDegree";
 import moveMonster from "./utils/physicsUtils/moveMonster";
 import makeReflectionAngle from "./utils/physicsUtils/makeReflectionAngle";
-import { translateEntitiesY } from "./utils/physicsUtils/translateEntity";
 import playAudio from "./utils/playAudio";
 import { swipe } from "../assets/audio";
 import { disPatchInteractionWithItem } from "./utils/physicsUtils/checkBoundary";
 import translateAllEntitiesX from "./utils/physicsUtils/translateMapX";
-import { settingBoss } from "./utils/physicsUtils/settingBoss";
+import { setGuideMissileUnderHPTen } from "./utils/physicsUtils/bossUtills/settingBoss";
+import translateAllEntitiesY from "./utils/physicsUtils/translateMapY";
 
 const WINDOW_HEIGHT = Dimensions.get("window").height;
-const WINDOW_WIDTH = Dimensions.get("window").width;
 const FLOOR_WIDTH = 32;
 const GAME_HEIGHT = WINDOW_HEIGHT - FLOOR_WIDTH;
-const GAME_WIDTH = WINDOW_WIDTH - FLOOR_WIDTH;
 
 let translateMapY = false;
 let translateMapX = false;
-let movedHeight = 0;
+const movedHeight = 0;
 const movedWidth = 0;
 
 export default function usePhysics(entities, { touches, dispatch }) {
   const { engine, world } = entities.physics;
   const player = entities.player.body;
-  const { stage, mapInfo, initialRotation } = entities;
+  const { stage, initialRotation } = entities;
   const monsterNumber = entityInfo[stage].monster.number;
   const monsterArray = Array.from({ length: monsterNumber }, (_, i) => i);
-  const itemNumber = entityInfo[stage].item.number;
-  const itemArray = Array.from({ length: itemNumber }, (_, i) => i);
-  const blockNumber = entityInfo[stage].block.renderEntity;
-  const blockArray = Array.from({ length: blockNumber }, (_, i) => i);
-  const firstBlocksNum = entityInfo[stage].block.firstEntity;
-  const firstBlockArray = Array.from({ length: firstBlocksNum }, (_, i) => i);
-  const specialItemNumber = entityInfo[stage].special.number;
-  const specialItemArray = Array.from(
-    { length: specialItemNumber },
-    (_, i) => i,
-  );
   const attackNumber = entityInfo[stage].attack.number;
   const attackArray = Array.from({ length: attackNumber }, (_, i) => i);
-  const flagNumber = entityInfo[stage].flag.number;
 
   Matter.Engine.update(engine);
   engine.timing.delta = 1 / 60;
@@ -71,95 +57,17 @@ export default function usePhysics(entities, { touches, dispatch }) {
     entities,
   });
 
-  if (translateMapY) {
-    if (!movedHeight) {
-      dispatch({ type: "move_page" });
-    }
-    if (movedHeight < GAME_HEIGHT) {
-      movedHeight += 10;
-
-      Matter.Body.setVelocity(entities.player.body, {
-        x: 0,
-        y: 0,
-      });
-
-      Matter.Body.translate(entities.player.body, { x: 0, y: 10 });
-
-      translateEntitiesY(blockNumber, 10, "block", entities, Matter);
-      translateEntitiesY(flagNumber, 10, "flag", entities, Matter);
-      translateEntitiesY(monsterNumber, 10, "monster", entities, Matter);
-
-      Matter.Body.translate(entities.boss1.body, {
-        x: 0,
-        y: 10,
-      });
-    } else {
-      firstBlockArray.forEach((entityNum) => {
-        Matter.Body.setPosition(entities[`block${entityNum + 1}`].body, {
-          x: mapInfo.block[entityNum + 38].position.x,
-          y: mapInfo.block[entityNum + 38].position.y + movedHeight,
-        });
-      });
-
-      Matter.Body.translate(entities.player.body, { x: 0, y: 10 });
-
-      dispatch({ type: "complete_move_upper", payload: movedHeight });
-
-      itemArray.forEach((num) => {
-        mapInfo.item[num + 1].position.y += movedHeight;
-      });
-
-      specialItemArray.forEach((num) => {
-        mapInfo.special[num + 1].position.y += movedHeight;
-      });
-
-      entities.round += 1;
-      translateMapY = false;
-      entities.translatedInfo.y += movedHeight;
-      movedHeight = 0;
-
-      settingBoss({ entities, Matter, bossRound:4 });
-    }
-  }
+  translateAllEntitiesY({
+    translateMapY,
+    movedHeight,
+    Matter,
+    entities,
+    dispatch,
+  });
 
   if (entities.round === 4) {
-    const bossHP = entityInfo[stage].boss.specifics[1].HP;
-    const monsterSpecific = entityInfo[stage].monster.specifics;
-
-    if (bossHP === 10) {
-      Matter.Body.setPosition(entities.monster3.body, {
-        x: entities.boss1.body.position.x,
-        y: entities.boss1.body.position.y,
-      });
-      entityInfo[stage].monster.specifics[3].guideMissile = {
-        x: (player.position.x - entities.boss1.body.position.x) / 40,
-        y: (player.position.y - entities.boss1.body.position.y) / 40,
-      };
-    }
-
-    const MonsterPositionTimer = entityInfo[stage].monster.specifics[1].moved;
-
-    if (
-      MonsterPositionTimer > (GAME_HEIGHT / 3) * 1 &&
-      monsterSpecific[2].onPosition === false
-    ) {
-      monsterSpecific[2].onPosition = true;
-      Matter.Body.setPosition(entities.monster2.body, {
-        x: entities.boss1.body.bounds.min.x,
-        y: entities.boss1.body.bounds.max.y,
-      });
-    }
-
-    if (
-      MonsterPositionTimer > (GAME_HEIGHT / 3) * 2 &&
-      monsterSpecific[5].onPosition === false
-    ) {
-      monsterSpecific[5].onPosition = true;
-      Matter.Body.setPosition(entities.monster5.body, {
-        x: entities.boss1.body.bounds.max.x,
-        y: entities.boss1.body.bounds.max.y,
-      });
-    }
+    setGuideMissileUnderHPTen({ Matter, entities });
+    startThrowMonsterTimely({ Matter, entities });
 
     monsterArray.forEach((entityNum) => {
       const specifics = entityInfo[stage].monster.specifics[entityNum + 1];
@@ -180,6 +88,7 @@ export default function usePhysics(entities, { touches, dispatch }) {
           });
         }
       }
+
       if (specifics.guideMissile) {
         Matter.Body.translate(entities[`monster${entityNum + 1}`].body, {
           x: specifics.guideMissile.x,
